Commit 2
Changed the way to remove from a index, because the code stopped working with mutiple indexes. If I want to make the check boxes work the way I want I also will have todo the same with it. It still does the correct tasks though. However checkbox is added and will change with button an active listen.